Updated .gitignore. Removed leftover from old repo, bare essentials now
Looks to me that the .gitignore file was the one generated by a React Native starter project with some additions at the bottom.

Later I will add a sample React Native project in a subfolder. That folder will contain it's own .gitignore file which will probably be more or less like this file was.