ExampleApp should now correctly include react-native-hugecharts as dep.
The ExampleApp should now correctly include react-native-hugecharts as a dependency.

It was a learning experience to get the subfolder ExampleApp/ to use react-native-hugecharts as a dependency correctly. Primarily what I learned along the way was using "file:../" as a dependency source in ExampleApp/package.json. It essentially says: my dependency is in the folder above. The second thing I learned was .npmignore files. Without it I was getting cryptic dependency conflicts but the reason was that the ExampleApp/ folder was included creating a recursive cycle of dependencies.

I deleted ExampleApp/index.js and ExampleApp/page.html as they were temporary files to get the sample app going. No longer needed as they are now fetched correctly from the root.

In package.json I include react and react-native as peer dependencies. The versions I choose were the versions I'm using in my production app which will be the first app to eventually use this library. I wanted to have things as similar as possible to begin with and then I’ll hopefully upgrade to a more recent version later.